Fix wrong test
Separation of interests.
