+++ package.json
@@ -85,9 +85,7 @@
     "jest": "^24.9.0",
     "ts-jest": "^24.1.0",
     "tslint": "^5.12.1",
-    "typescript": "^3.3.1"
-  },
-  "dependencies": {
+    "typescript": "^3.3.1",
     "lodash": "^4.17.15"
   }
 }
