+++ src/utils/file.ts
@@ -1,24 +1,73 @@
-import * as fs from 'fs';
-import * as util from 'util';
 import * as path from 'path';
+import { promises } from 'fs';
 
-import { rootFile, basePath } from '../config';
+import { deleteExtname } from './word';
 
-export const createWikiDirectory = () => {
-  return util.promisify(fs.mkdir)(basePath, { recursive: false });
+const frontMatterLineRegExp = new RegExp(/---/g);
+
+export const createWikiDirectory = (path: string) => {
+  return promises.mkdir(path, { recursive: false });
 };
 
 export const createIndexFile = async (): Promise<string> =>
   await createFile('index.md');
 
-export const createFile = async (name: string): Promise<string> => {
-  const filePath = path.join(basePath, name);
-
+export const createFile = async (filePath: string): Promise<string> => {
   try {
-    await util.promisify(fs.stat)(filePath);
+    await promises.stat(filePath);
   } catch (err) {
-    await util.promisify(fs.writeFile)(filePath, '');
+    await promises.writeFile(filePath, '');
   }
 
   return filePath;
+};
+
+export const copyFile = async (origin: string, target: string) => {
+  const originText = (await promises.readFile(origin)).toString();
+
+  const title = path.basename(target, '.md');
+
+  let newText: string;
+  let frontMatter: string;
+
+  try {
+    await promises.stat(target);
+
+    const targetContent = (await promises.readFile(target)).toString();
+
+    frontMatter = extractFrontMatter(targetContent);
+
+    newText = frontMatter + '\n\n' + deleteExtname(originText);
+
+    if (newText === targetContent) {
+      return;
+    }
+  } catch (err) {
+    frontMatter = `---
+title: ${title}
+---`;
+
+    newText = frontMatter + '\n\n' + deleteExtname(originText);
+  }
+
+  await promises.writeFile(target, newText);
+};
+
+export const extractFrontMatter = (content: string) => {
+  const boundaries = [];
+
+  while (true) {
+    const match = frontMatterLineRegExp.exec(content);
+    if (!match) {
+      break;
+    }
+
+    boundaries.push(match.index);
+  }
+
+  if (boundaries.length < 2 || boundaries[0] !== 0) {
+    throw new Error('Front Matter is not exist');
+  }
+
+  return content.slice(boundaries[0], boundaries[1] + '---'.length);
 };
