+++ src/utils/word.test.ts
@@ -1,11 +1,10 @@
-import { extractLinkName, containsInLink } from './word';
+import { extractLinkName, containsInLink, deleteExtname } from './word';
 
 test('extract word', () => {
   expect(extractLinkName('[Docker](Docker)')).toBe('Docker');
   expect(extractLinkName('[Text with space](text with space)'))
     .toBe('text-with-space');
 });
-
 
 test('containsInLink', () => {
   const text = '## How to install [Docker](Docker) on [macOS X](macOS X) using HomeBrew';
@@ -14,4 +13,12 @@ test('containsInLink', () => {
   expect(containsInLink(text, 18, 32)).toBe('[Docker](Docker)');
   expect(containsInLink(text, 18, 34)).toBe('[Docker](Docker)');
   expect(containsInLink(text, 38, 39)).toBe('[macOS X](macOS X)');
+});
+
+test('deleteExtname', () => {
+  expect(
+    deleteExtname(
+      'Install [Docker](Docker.md) with [Homebrew](Homebrew.md) test.md'
+    )
+  ).toBe('Install [Docker](Docker) with [Homebrew](Homebrew) test.md');
 });
